Server/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
tqdm
pydantic
python-multipart
httpx
python-dotenv
